+++ hotel-booking-backend/src/index.ts
@@ -265,7 +265,7 @@ app.use(
     origin: corsOriginHandler,
     credentials: true,
     optionsSuccessStatus: 204,
-    methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
+    methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"],
     allowedHeaders: [
       "Content-Type",
       "Authorization",
@@ -281,7 +281,7 @@ app.options(
     origin: corsOriginHandler,
     credentials: true,
     optionsSuccessStatus: 204,
-    methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
+    methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"],
     allowedHeaders: [
       "Content-Type",
       "Authorization",